added separate folder for Ships
