+++ src/components/Games/Ships/ships.js
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef } from "react";
-import Navbar from "../Navbar/Navbar";
-import GoBack from "../GoBack/GoBack";
-import ShipsComputerArea from "./shipCompArea";
+import Navbar from "../../Navbar/Navbar";
+import GoBack from "../../GoBack/GoBack";
+import ShipsComputerArea from "./computerArea";
 
 const ShipsGame = () => {
   return (
