+++ src/pages/Games/Games.js
@@ -8,10 +8,10 @@ const Games = () => {
       <Navbar />
       <GoBack />
       <h1 id="games-header">Games</h1>
-      <NavLink className="game-link-button" to="/games/rps">
+      <NavLink className="game-link-button" to="/games/Ships/rps">
         RPS
       </NavLink>
-      <NavLink className="game-link-button" to="/games/ships">
+      <NavLink className="game-link-button" to="/games/Ships/ships">
         SHIPS
       </NavLink>
     </article>
